fix from Quenten

+++ src/Devices/Bed.cpp
@@ -21,8 +21,8 @@ Bed::~Bed() {
 std::string Bed::handleSensorUpdate(XmlReader * xml_r) {
     /* init variables to be used */
     std::string destination;
-    std::vector<double> sentStatus;
-    std::vector<double> sendStatus;
+    std::vector<float> sentStatus;
+    std::vector<float> sendStatus;
 
     destination = xml_r->getSenderName();
     sentStatus = xml_r->getData();
@@ -71,8 +71,8 @@ std::string Bed::handleSensorUpdate(XmlReader * xml_r) {
 std::string Bed::handleWebsiteUpdate(XmlReader * xml_r) {
     /* init variables to be used */
     std::string destination;
-    std::vector<double> sentStatus;
-    std::vector<double> sendStatus;
+    std::vector<float> sentStatus;
+    std::vector<float> sendStatus;
     //std::cout << "reading destination" << std::endl;
     destination = xml_r->getClientName();
     //std::cout << "getting data" << std::endl;
