+++ src/Devices/Column.cpp
@@ -18,8 +18,8 @@ Column::~Column(){
 
 std::string Column::handleSensorUpdate(XmlReader* xml_r) {
     std::string destination;
-    std::vector<double> sentStatus;
-    std::vector<double> sendStatus;
+    std::vector<float> sentStatus;
+    std::vector<float> sendStatus;
 
     destination = xml_r->getSenderName();
     sentStatus = xml_r->getData();
@@ -54,8 +54,8 @@ std::string Column::handleSensorUpdate(XmlReader* xml_r) {
 std::string Column::handleWebsiteUpdate(XmlReader * xml_r) {
     /* init variables to be used */
     std::string destination;
-    std::vector<double> sentStatus;
-    std::vector<double> sendStatus;
+    std::vector<float> sentStatus;
+    std::vector<float> sendStatus;
     //std::cout << "reading destination" << std::endl;
     destination = xml_r->getClientName();
     //std::cout << "getting data" << std::endl;
